🚚 move parseFlags and parseArgs into Command class
